# bpo-31993: do not allocate large temporary buffers in pickle dump