… Lib/test/pickletester.py
… Lib/test/test_pickletools.py
… Misc/NEWS.d/next/Library/2017-11-10-00-05-08.bpo-31993.-OMNg8.rst