@ogrisel 2017-11-09
Thanks for the quick review.

I think I addressed your comments. I am not sure how to write the entry in `NEWS.d`. I think this PR is useful as it is and we should not wait for the C version. Shall I add an entry for the Python _Pickler alone in the NEWS file for 3.7.0a2?

I would appreciate it if someone else could do the C version.