@pitrou 2017-11-09
Yes, you should add a NEWS entry.  You can (and should) use the ["blurb" utility](https://pypi.python.org/pypi/blurb) for that.

I think it's fine to leave the C version for later or to someone else if you're not comfortable with that.