@ogrisel 2017-11-09
Done. Feel free to edit it directly if you want to change anything.