@serhiy-storchaka 2017-11-10
I'm not sure about 5d463159bfecd5fafb4745e63c0f24c6d0815fdf. opcode and size_header are concatenated in any case. If there is any benefit from this change, how can it be explained?