@serhiy-storchaka 2017-11-10
Ah, you already have reverted this change. The final variant looks pretty simple.