@serhiy-storchaka 2017-11-10
Don't squash them. It is more convenient to make a review if intermediate commits are not squashed. All them will be squashed when be merged in master.