@ogrisel 2017-11-11
@serhiy-storchaka approved this PR but I subsequently included a implementation for the C version that has not been reviewed yet. The "awaiting-merge" label should be removed and replaced by "awaiting-review".