@ogrisel 2017-11-12
Do you want me to convert [this script](https://gist.github.com/ogrisel/0e7b3282c84ae4a581f3b9ec1d84b45a) into a new non-regression test that ensures that future changes to the pickle module will no break the no-copy semantics dump for large bytes & str?