@ogrisel 2017-11-12
The CI / github integration seems to be confused by the [ci skip]. I will push again my last 2 commits in reverse order.