@ogrisel 2017-11-12
I have made the requested changes; please review again.

In particular:
- Large objects are no longer wrapped in frames.
- I have implemented the "flush after frame commit" behavior for the C pickler and tried make the Python pickler behave consistently (check the new test: `test_framed_write_sizes`.

As a consequence, it means that the C implementation of protocol 3 makes a single write (at the end of dump) while protocol 4 can issue many calls to write (`total_size / 64 kB` approximately). Maybe we should consider adding an option to control the target frame size in protocol 4. Also, we might want to make it larger by default to avoid introducing any performance regression on typical workloads.