@ogrisel 2017-11-13
I rebased because the CI status was stalled again.