@ogrisel 2017-11-13
Thanks for the review. I am working on other issues today and might be a bit busy on other things in the coming days but I plan to address them as soon as possible.