@ogrisel 2017-11-17
I think I addressed all the comments. Do you want me to add a memory test to protect against any future regression in the new no-copy semantics of the `dump` methods?