@pitrou 2017-11-17
> Do you want me to add a memory test to protect against any future regression in the new no-copy semantics of the dump methods?

This is just an optimization, not a semantic guarantee. Besides, it's quite tricky to measure memory consumption in a reliable and cross-platform way.