@ogrisel 2017-11-17
Also: shall we make `frame_size_target` an additional argument to the `Pickler` class? Now that we use it to trigger flushes systematically for the C pickler this can have a significant impact on the pickling speed.