@pitrou 2017-11-17
> Also: shall we make frame_size_target an additional argument to the Pickler class?

I would leave that for a separate PR. Perhaps you will want to benchmark such a change first on various kinds of IO objects.