@ogrisel 2017-11-21
Sorry for the lack of reply. I will work on running some benchmark soon.