@ogrisel 2017-11-24
Here are some benchmark results (for the C implementation). As expected calling write more often causes a performance regression.

On master:

```
(py37) ogrisel@ici:~/code/cpython$ ./python ~/tmp/bench_pickle.py 
# protocol 2, output: output.pkl
median time: 0.247s
# protocol 3, output: output.pkl
median time: 0.249s
# protocol 4, output: output.pkl
median time: 0.25s
# protocol 2, output: /dev/null
median time: 0.142s
# protocol 3, output: /dev/null
median time: 0.144s
# protocol 4, output: /dev/null
median time: 0.15s
last pickle size: 53.330MB
```

On this branch:

```
(py37) ogrisel@ici:~/code/cpython$ ./python ~/tmp/bench_pickle.py 
# protocol 2, output: output.pkl
median time: 0.286s
# protocol 3, output: output.pkl
median time: 0.284s
# protocol 4, output: output.pkl
median time: 0.285s
# protocol 2, output: /dev/null
median time: 0.175s
# protocol 3, output: /dev/null
median time: 0.174s
# protocol 4, output: /dev/null
median time: 0.174s
last pickle size: 53.330MB
```