@ogrisel 2017-11-24
Here is the code of the benchmark:

```python
from time import time
import pickle
from statistics import median

# Nested datastructure with many small objects that should result in a
# pickle significantly larger than the target frame size:
objects = [{str(i): (bytes(i % 1000), i) for i in range(int(1e5))}]

ondisk_filename = 'output.pkl'

for filename in [ondisk_filename, '/dev/null']:
    for proto in [2, 3, 4]:
        print(f"# protocol {proto}, output: {filename}")

        timings = []
        for i in range(20):
            with open(filename, 'wb') as f:
                tic = time()
                pickle.dump(objects, f)
                toc = time()
                timings.append(toc - tic)
        print(f"median time: {median(timings):0.3}s")


with open(ondisk_filename, 'rb') as f:
    print(f"last pickle size: {len(f.read()) / 1e6:0.3f}MB")
```