@pitrou 2017-11-24
Interesting. What if you make the frame size larger?