@ogrisel 2017-11-25
There is something fishy with the results of my benchmark: for protocol 2 and 3, there is no framing enabled, therefore the pickler still only writes once at the end of the dump. When I increase the frame size the difference is smaller but I suspect that I have other things going on my machine because I get widely different results despite taking the median time of 20 runs for each measurement.