@pitrou 2017-11-25
Two suggestions:
* make the benchmark longer, to avoid effects of CPU frequency scaling
* pin the benchmark to a particular CPU core (for example using `taskset` on Linux)