@pitrou 2017-11-25
I tried the benchmark on my machine and I also got a lot of variation. Sometimes the unpatched version is faster, sometimes the patched version is faster. If you want to be perfectionist, perhaps you want to try @vstinner's `perf` tool: https://perf.readthedocs.io/en/latest/