@vstinner 2017-11-27
I'm disapointed that @ogrisel posted a benchmark without the standard deviation. What happened to you Olivier? :-)
http://perf.readthedocs.io/en/latest/analyze.html#minimum-vs-average