@vstinner 2017-11-27
My advices to get reproductible results:
http://perf.readthedocs.io/en/latest/run_benchmark.html#how-to-get-reproductible-benchmark-results