@ogrisel 2017-11-27
I could have bootstrapped the median to build confidence intervals but it's not (yet?) part of the `statistics` module of the standard library :)

Anyway I used `python3 -m perf tune` to set my system in "performance" mode, I use taskset to run the bench on CPU 0 only, and disabled the GC and now I get stable benchmarks:

- On master:

```
# protocol 2, output: output.pkl
median time: 0.336s
# protocol 3, output: output.pkl
median time: 0.338s
# protocol 4, output: output.pkl
median time: 0.336s
# protocol 2, output: /dev/null
median time: 0.191s
# protocol 3, output: /dev/null
median time: 0.189s
# protocol 4, output: /dev/null
median time: 0.189s
last pickle size: 53.330MB
```

- On this branch:

```
$ taskset -c 0 ./python ~/tmp/bench_pickle.py 
# protocol 2, output: output.pkl
median time: 0.333s
# protocol 3, output: output.pkl
median time: 0.334s
# protocol 4, output: output.pkl
median time: 0.334s
# protocol 2, output: /dev/null
median time: 0.185s
# protocol 3, output: /dev/null
median time: 0.188s
# protocol 4, output: /dev/null
median time: 0.188s
last pickle size: 53.330MB
```

So there is no performance regression on large collections of small objects.