@ogrisel 2017-12-13
I ran a benchmark with 63 MB of 63kB random bytes objects and there is no performance regression when using getvalue in this PR vs master:

- master:

```
# protocol 4, output: output.pkl
median time: 0.23s
# protocol 4, output: /dev/null
median time: 0.0262s
last pickle size: 63.009MB
```

- this PR with `getvalue` instead of `getbuffer` 

```
# protocol 4, output: output.pkl
median time: 0.23s
# protocol 4, output: /dev/null
median time: 0.0256s
last pickle size: 63.009MB
```

with `getbuffer` it's the same.